refactor: WebSocket SessionContext 공용화 (#51)
