+++ src/main/java/config/WebSocketConfig.java
@@ -1,11 +1,11 @@
-package lobby.ws;
+package config;
 
 import javax.servlet.http.HttpSession;
 import javax.websocket.HandshakeResponse;
 import javax.websocket.server.HandshakeRequest;
 import javax.websocket.server.ServerEndpointConfig;
 
-public class LobbyWebSocketConfig extends ServerEndpointConfig.Configurator {
+public class WebSocketConfig extends ServerEndpointConfig.Configurator {
 
 	@Override
 	public void modifyHandshake(ServerEndpointConfig sec,
