+++ src/main/java/session/ContextDTO.java
@@ -1,4 +1,4 @@
-package lobby.dto;
+package session;
 
 import javax.servlet.http.HttpSession;
