+++ src/main/java/session/Role.java
@@ -1,4 +1,4 @@
-package lobby.dto;
+package session;
 
 public enum Role {
 	USER, // 로그인 회원
